Move fixture to separate header.
